% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaVelocyto-package.R
\name{delta}
\alias{delta}
\title{compare two slope of unspliced vs spliced moments in a gene of two samples}
\usage{
delta(ladata1, ladata2, gene)
}
\description{
compare two slope of unspliced vs spliced moments in a gene of two samples
}
