\name{DeltaVelocyto-package}
\alias{DeltaVelocyto-package}
\alias{DeltaVelocyto}
\docType{package}
\title{
\packageTitle{DeltaVelocyto}
}
\description{
\packageDescription{DeltaVelocyto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DeltaVelocyto}
\packageIndices{DeltaVelocyto}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DeltaVelocyto}

Maintainer: \packageMaintainer{DeltaVelocyto}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
