% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaVelocyto-package.R
\name{filter_and_norm}
\alias{filter_and_norm}
\title{Select the top (most variable) n genes with a minimum of x shared counts across cells}
\usage{
filter_and_norm(ladata, n = 1000, min_counts = 20)
}
\description{
Select the top (most variable) n genes with a minimum of x shared counts across cells
}
