% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaVelocyto-package.R
\name{lm}
\alias{lm}
\title{Find the linear model of spliced and unspliced moments in a gene}
\usage{
lm(ladata, gene)
}
\description{
Find the linear model of spliced and unspliced moments in a gene
}
